Mais algumas mudanças
Continuei a implementar a IA atualizada do Obscuro (talvez tenha quebrado algo).
Algumas mudanças na cena da ADM, trazendo o dano em área do obscuro dentro da área do altar norte.

+++ 6_Semestre/Assets/PlayerStats.cs
@@ -101,7 +101,7 @@ public class PlayerStats : MonoBehaviour
         fadeScript.StartCoroutine(fadeScript.Fade(3f));
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentInsanity -= damage;
         postprocessingControl.espiritualEnergy = currentInsanity;
